Let there be ACL!
git-svn-id: http://joomlacode.org/svn/joomla/development/trunk@13031 6f6e1ebd-4c2b-0410-823f-f34bde69bce9

+++ administrator/components/com_content/views/articles/view.html.php
@@ -45,30 +45,38 @@ class ContentViewArticles extends JView
 
 	/**
 	 * Display the toolbar
-	 *
-	 * @access	private
 	 */
 	protected function _setToolbar()
 	{
-		$state = $this->get('State');
+		$user	= JFactory::getUser();
+		$state	= $this->get('State');
+
 		JToolBarHelper::title(JText::_('Content_Articles_Title'), 'article.png');
-		JToolBarHelper::custom('article.edit', 'new.png', 'new_f2.png', 'New', false);	
-		JToolBarHelper::custom('article.edit', 'edit.png', 'edit_f2.png', 'Edit', true);
+		if ($user->authorise('core.create', 'com_content')) {
+			JToolBarHelper::custom('article.edit', 'new.png', 'new_f2.png', 'New', false);
+		}
+		if ($user->authorise('core.edit', 'com_content')) {
+			JToolBarHelper::custom('article.edit', 'edit.png', 'edit_f2.png', 'Edit', true);
+		}
 		JToolBarHelper::divider();
-		JToolBarHelper::custom('articles.publish', 'publish.png', 'publish_f2.png', 'Publish', true);
-		JToolBarHelper::custom('articles.unpublish', 'unpublish.png', 'unpublish_f2.png', 'Unpublish', true);
-		if ($state->get('filter.published') != -1) {
-			JToolBarHelper::archiveList('articles.archive');
-		}		
-		if ($state->get('filter.published') == -2) {
+		if ($user->authorise('core.edit.state', 'com_content')) {
+			JToolBarHelper::custom('articles.publish', 'publish.png', 'publish_f2.png', 'Publish', true);
+			JToolBarHelper::custom('articles.unpublish', 'unpublish.png', 'unpublish_f2.png', 'Unpublish', true);
+			if ($state->get('filter.published') != -1) {
+				JToolBarHelper::archiveList('articles.archive');
+			}
+		}
+		if ($state->get('filter.published') == -2 && $user->authorise('core.delete', 'com_content')) {
 			JToolBarHelper::deleteList('', 'articles.delete');
 		}
-		else {
+		else if ($user->authorise('core.edit.state', 'com_content')) {
 			JToolBarHelper::trash('articles.trash');
 		}
+		if ($user->authorise('core.admin', 'com_content')) {
+			JToolBarHelper::divider();
+			JToolBarHelper::preferences('com_content');
+		}
 		JToolBarHelper::divider();
-		JToolBarHelper::preferences('com_content');
-		JToolBarHelper::divider();		
 		JToolBarHelper::help('screen.content.articles');
 	}
 }
